backend/requirements.txt
FastAPI>=0.110.0
uvicorn[standard]>=0.27.0
sentence-transformers>=2.6.0
qdrant-client>=1.8.0
psycopg2-binary>=2.9.0
slowapi>=0.1.9
python-dotenv>=1.0.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
